Облегченный conftest. Разделил фикстуры по типу внешнего сервиса и перенес в fuxtures/
